scripts/rock_download/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.3.0